fixup! IBX-6827: Ranges generators
