+++ src/contracts/Repository/Values/Content/Query/Aggregation/DateMetadataRangeAggregation.php
@@ -8,6 +8,9 @@ declare(strict_types=1);
 
 namespace Ibexa\Contracts\Core\Repository\Values\Content\Query\Aggregation;
 
+use Ibexa\Contracts\Core\Repository\Values\Content\Query\Aggregation\Ranges\RangesGeneratorInterface;
+use Traversable;
+
 final class DateMetadataRangeAggregation extends AbstractRangeAggregation
 {
     public const MODIFIED = 'modified';
@@ -26,6 +29,19 @@ final class DateMetadataRangeAggregation extends AbstractRangeAggregation
     public function getType(): string
     {
         return $this->type;
+    }
+
+    public static function fromGenerator(
+        string $name,
+        string $type,
+        RangesGeneratorInterface $generator
+    ): self {
+        $ranges = $generator->generate();
+        if ($ranges instanceof Traversable) {
+            $ranges = iterator_to_array($ranges);
+        }
+
+        return new self($name, $type, $ranges);
     }
 }
